Rename files and add Home component.
