+++ step-release-vis/src/app/app.module.ts
@@ -2,15 +2,17 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
-import { SvgGridComponent } from './components/grid/svg-grid/svg-grid.component';
+import { SvgGridComponent } from './components/grid/svg/svg';
 import { AppRoutingModule } from './app-routing.module';
-import { CanvasGridComponent } from './components/grid/canvas-grid/canvas-grid.component';
+import { CanvasGridComponent } from './components/grid/canvas/canvas';
+import { HomeComponent } from './components/home/home';
 
 @NgModule({
   declarations: [
     AppComponent,
     SvgGridComponent,
-    CanvasGridComponent
+    CanvasGridComponent,
+    HomeComponent
   ],
   imports: [
     BrowserModule,
