--- step-release-vis/src/app/components/grid/svg-grid/svg-grid.component.html
@@ -1,5 +1,0 @@
-<!DOCTYPE html>
-<html>
-<body id="container">
-</body>
-</html>
